+++ src/observe.js
@@ -160,6 +160,8 @@
                     watch.originalObject = domElement.outerHTML;
                 } else {
                     watch.originalObject = _string.empty;
+
+                    fireCustomTrigger( watch.options.onRemove, watch.domElementId );
                 }
             }
 
@@ -291,6 +293,7 @@
         options.onChange = getDefaultFunction( options.onChange, null );
         options.onPropertyChange = getDefaultFunction( options.onPropertyChange, null );
         options.onCancel = getDefaultFunction( options.onCancel, null );
+        options.onRemove = getDefaultFunction( options.onRemove, null );
 
         return options;
     }