+++ src/observe.js
@@ -189,10 +189,12 @@
                     var oldValue = getObjectFromString( cachedObject ).result,
                         newValue = getObjectFromString( originalObjectJson ).result;
 
-                    compareWatchObject( oldValue, newValue, watch );
+                    if ( !isDefinedArray( oldValue ) && !isDefinedArray( newValue ) ) {
+                        compareWatchObject( oldValue, newValue, watch );
 
-                    if ( isDefinedFunction( watch.options.onPropertyChange ) && !isDefinedArray( oldValue ) ) {
-                        compareWatchObjectProperties( oldValue, newValue, watch );
+                        if ( isDefinedFunction( watch.options.onPropertyChange ) ) {
+                            compareWatchObjectProperties( oldValue, newValue, watch );
+                        }
                     }
                 }